reorg code: move interface reference to implementation
