+++ pkg/provider/azure_instances_v1.go
@@ -32,6 +32,8 @@ import (
 	"sigs.k8s.io/cloud-provider-azure/pkg/consts"
 )
 
+var _ cloudprovider.Instances = (*Cloud)(nil)
+
 const (
 	vmPowerStatePrefix       = "PowerState/"
 	vmPowerStateStopped      = "stopped"
@@ -220,38 +222,6 @@ func (az *Cloud) InstanceExistsByProviderID(ctx context.Context, providerID stri
 	return true, nil
 }
 
-// InstanceExists returns true if the instance for the given node exists according to the cloud provider.
-// Use the node.name or node.spec.providerID field to find the node in the cloud provider.
-func (az *Cloud) InstanceExists(ctx context.Context, node *v1.Node) (bool, error) {
-	if node == nil {
-		return false, nil
-	}
-	unmanaged, err := az.IsNodeUnmanaged(node.Name)
-	if err != nil {
-		return false, err
-	}
-	if unmanaged {
-		klog.V(4).Infof("InstanceExists: omitting unmanaged node %q", node.Name)
-		return true, nil
-	}
-
-	providerID := node.Spec.ProviderID
-	if providerID == "" {
-		var err error
-		providerID, err = cloudprovider.GetInstanceProviderID(ctx, az, types.NodeName(node.Name))
-		if err != nil {
-			if strings.Contains(err.Error(), cloudprovider.InstanceNotFound.Error()) {
-				return false, nil
-			}
-
-			klog.Errorf("InstanceExists: failed to get the provider ID by node name %s: %v", node.Name, err)
-			return false, err
-		}
-	}
-
-	return az.InstanceExistsByProviderID(ctx, providerID)
-}
-
 // InstanceShutdownByProviderID returns true if the instance is in safe state to detach volumes
 func (az *Cloud) InstanceShutdownByProviderID(ctx context.Context, providerID string) (bool, error) {
 	if providerID == "" {
@@ -297,38 +267,6 @@ func (az *Cloud) InstanceShutdownByProviderID(ctx context.Context, providerID st
 	status := strings.ToLower(powerStatus)
 	provisioningSucceeded := strings.EqualFold(strings.ToLower(provisioningState), strings.ToLower(string(consts.ProvisioningStateSucceeded)))
 	return provisioningSucceeded && (status == vmPowerStateStopped || status == vmPowerStateDeallocated || status == vmPowerStateDeallocating), nil
-}
-
-// InstanceShutdown returns true if the instance is shutdown according to the cloud provider.
-// Use the node.name or node.spec.providerID field to find the node in the cloud provider.
-func (az *Cloud) InstanceShutdown(ctx context.Context, node *v1.Node) (bool, error) {
-	if node == nil {
-		return false, nil
-	}
-	unmanaged, err := az.IsNodeUnmanaged(node.Name)
-	if err != nil {
-		return false, err
-	}
-	if unmanaged {
-		klog.V(4).Infof("InstanceShutdown: omitting unmanaged node %q", node.Name)
-		return false, nil
-	}
-	providerID := node.Spec.ProviderID
-	if providerID == "" {
-		var err error
-		providerID, err = cloudprovider.GetInstanceProviderID(ctx, az, types.NodeName(node.Name))
-		if err != nil {
-			// Returns false, so the controller manager will continue to check InstanceExistsByProviderID().
-			if strings.Contains(err.Error(), cloudprovider.InstanceNotFound.Error()) {
-				return false, nil
-			}
-
-			klog.Errorf("InstanceShutdown: failed to get the provider ID by node name %s: %v", node.Name, err)
-			return false, err
-		}
-	}
-
-	return az.InstanceShutdownByProviderID(ctx, providerID)
 }
 
 func (az *Cloud) isCurrentInstance(name types.NodeName, metadataVMName string) (bool, error) {
@@ -502,59 +440,6 @@ func (az *Cloud) AddSSHKeyToAllInstances(ctx context.Context, user string, keyDa
 // On Azure this is the hostname, so we just return the hostname.
 func (az *Cloud) CurrentNodeName(ctx context.Context, hostname string) (types.NodeName, error) {
 	return types.NodeName(hostname), nil
-}
-
-// InstanceMetadata returns the instance's metadata. The values returned in InstanceMetadata are
-// translated into specific fields in the Node object on registration.
-// Use the node.name or node.spec.providerID field to find the node in the cloud provider.
-func (az *Cloud) InstanceMetadata(ctx context.Context, node *v1.Node) (*cloudprovider.InstanceMetadata, error) {
-	meta := cloudprovider.InstanceMetadata{}
-	if node == nil {
-		return &meta, nil
-	}
-	unmanaged, err := az.IsNodeUnmanaged(node.Name)
-	if err != nil {
-		return &meta, err
-	}
-	if unmanaged {
-		klog.V(4).Infof("InstanceMetadata: omitting unmanaged node %q", node.Name)
-		return &meta, nil
-	}
-
-	if node.Spec.ProviderID != "" {
-		meta.ProviderID = node.Spec.ProviderID
-	} else {
-		providerID, err := cloudprovider.GetInstanceProviderID(ctx, az, types.NodeName(node.Name))
-		if err != nil {
-			klog.Errorf("InstanceMetadata: failed to get the provider ID by node name %s: %v", node.Name, err)
-			return nil, err
-		}
-		meta.ProviderID = providerID
-	}
-
-	instanceType, err := az.InstanceType(ctx, types.NodeName(node.Name))
-	if err != nil {
-		klog.Errorf("InstanceMetadata: failed to get the instance type of %s: %v", node.Name, err)
-		return &cloudprovider.InstanceMetadata{}, err
-	}
-	meta.InstanceType = instanceType
-
-	nodeAddresses, err := az.NodeAddresses(ctx, types.NodeName(node.Name))
-	if err != nil {
-		klog.Errorf("InstanceMetadata: failed to get the node address of %s: %v", node.Name, err)
-		return &cloudprovider.InstanceMetadata{}, err
-	}
-	meta.NodeAddresses = nodeAddresses
-
-	zone, err := az.GetZoneByNodeName(ctx, types.NodeName(node.Name))
-	if err != nil {
-		klog.Errorf("InstanceMetadata: failed to get the node zone of %s: %v", node.Name, err)
-		return &cloudprovider.InstanceMetadata{}, err
-	}
-	meta.Zone = zone.FailureDomain
-	meta.Region = zone.Region
-
-	return &meta, nil
 }
 
 // mapNodeNameToVMName maps a k8s NodeName to an Azure VM Name
